src/backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-pinecone
langchain-community
sentence-transformers
pinecone-client
groq
scikit-learn
numpy
python-dotenv